requirements.txt
Django>=5.2,<5.3
Pillow>=9.0.0
python-dotenv>=1.0.0
pytz
whitenoise>=6.9.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
supervisor>=4.2.5
dj-database-url>=2.1.0
django-cors-headers>=4.3.1
sentry-sdk>=1.40.0
django-cleanup>=8.1.0